V.0.5.3.9.8_RoadMap_2: полное сохранение проекта (сессия 2026-09-05); актуализация версий до 0.5.3.9.8 (version.txt, pyproject, frontend/constants, README, ENIGMA_ROADMAP). Р-В (SpeechExposure, GC-DIALOGUE-01): SSOT exposure_radius в communication (лестница secret/whisper/normal/loud/shout/private; normal=6.0 — parity D1 с ACTION_PERCEPTION_RADIUS, loud=10.0 — D2); IntentEventAdapter D3: secret→whisper (адресат слышит вплотную), loud→public; DialogueMaterializer D4: солилоквий-сентинел → whisper + SSOT-радиус; WorkingMemory NPC_SPOKE: radius из SSOT (999-дефолт запрещён ADR-148); routes scene_state → единый приёмник GameLoop.save_scene_state (R2-В/S244, anti-writer G3, whitelist-семантика субсумирована by construction). Mypy-доводка 641→608: 33 мелочи закрыты без вмешательства в логику (аннотации __init__/update/_init_db/main, X|None-дефолты, no-any-return → bool()/float(), var-annotated словари, TYPE_CHECKING-совместимые импорты). Гейты: тесты Р-В micro 10/10, micro-suite 11/11, импорты ядра OK. Примечание: unstaged-патчи сессии восстановлены из байткод-анализа __pycache__ после случайного checkout — routes/working_memory_tick байт-в-байт, communication/adapter/materializer dis-эквивалентны (расхождение только в формулировках комментариев).

+++ backend/app/services/affective/affective_integrator.py
@@ -21,7 +21,7 @@ _SURPRISE_GAIN: float = 1.2
 
 def _compute_scar_rate(abs_error: float) -> float:
     """Вычисляет скорость формирования 'шрама памяти' на основе ошибки предсказания."""
-    return 0.1 + (0.4 * (abs_error**1.5))
+    return float(0.1 + (0.4 * (abs_error**1.5)))
 
 
 def integrate_affective_pressure(
